add upsert for htable

+++ include/Srl/TpTools.h
@@ -139,7 +139,7 @@ namespace Srl { namespace TpTools {
         {
             double val = TP == Type::FP32 ? value.pblock().fp32 : value.pblock().fp64;
 
-            if(SrlType<T>::type != TP && (val < min<T>() || val > max<T>())) {
+            if(SrlType<T>::type != TP && (val < min<T>() || val > (double)max<T>())) {
                 return false;
             }
